comment_app/requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.7
django-crispy-forms==2.0
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.4.0
sqlparse==0.4.3
